Motion Controller Settings
Added settings for the motion controller to be set when the program first connects to the motion controller. Later, these settings should be read in from an external settings file so that they can be modified if needed without recompiling the program.
I also cleaned up the comments a bit and changed the name of the 'connect to controller' UI button from 'pushbutton_2' to 'connectController'.